Transferred CourseSetting page component to layouts
